Move component outside of core
