+++ cmd/agent/subcommands/run/command.go
@@ -31,8 +31,8 @@ import (
 	internalsettings "github.com/DataDog/datadog-agent/cmd/agent/subcommands/run/internal/settings"
 	agenttelemetry "github.com/DataDog/datadog-agent/comp/core/agenttelemetry/def"
 	agenttelemetryfx "github.com/DataDog/datadog-agent/comp/core/agenttelemetry/fx"
-	ownerdetectionfx "github.com/DataDog/datadog-agent/comp/core/ownerdetection/fx"
 	haagentfx "github.com/DataDog/datadog-agent/comp/haagent/fx"
+	ownerdetectionfx "github.com/DataDog/datadog-agent/comp/ownerdetection/fx"
 	compressionfx "github.com/DataDog/datadog-agent/comp/serializer/compression/fx"
 
 	// checks implemented as components
