+++ comp/ownerdetection/def/component.go
